feat: update permission service
