+++ Study/Emily/ReactNativeStudy/README.md
@@ -1,0 +1,30 @@
+# React Native Study
+
+## 1주차 과제
+
+### 1 네비게이션 이용하기
+
+### 2 박스 컴포넌트 사용하기
+
+### 3 전역 상태로 카운트 관리하기
+
+### 4 카카오 로그인 구현하기
+
+## 2주차 과제
+
+### 1 권한 구현하기
+
+<https://www.npmjs.com/package/react-native-permissions>
+
+### 2 카메라 촬영 & 이미지 선택 기능 구현하기
+
+<https://www.npmjs.com/package/react-native-image-crop-picker?activeTab=readme>
+
+### 3 스플래쉬 구현하기
+
+<https://www.npmjs.com/package/react-native-splash-screen>
+
+스플래쉬 아이콘 <https://www.flaticon.com/packs/april-fools-day-32>
+<https://velog.io/@2ast/React-Native-android-app-icon%EA%B3%BC-splash-screen-%EC%A0%81%EC%9A%A9%ED%95%98%EA%B8%B0>
+<https://hotpot.ai/>
+<https://github.com/crazycodeboy/react-native-splash-screen/issues/606#issuecomment-1396914012>
