+++ Study/Emily/ReactNativeStudy/src/services/Permission/PermissionService.ts
@@ -1,20 +1,51 @@
-import { Permission, check, request, RESULTS } from 'react-native-permissions';
+import { PERMISSIONS, Permission, RESULTS, checkMultiple, requestMultiple } from 'react-native-permissions';
+import { Permissions } from '../../models/Enums/Permissions';
 import { OS } from '../../models/Enums/OS';
+import { Platform } from 'react-native';
 
-export class PermissionService {
-  constructor() {}
+export const requestPermissions = async () => {
+  try {
+    const permissions = Object.values(Permissions);
 
-  key: Permission | undefined;
+    const permissionKeys = permissions.map((p) => PermissionFactory(p)) as Permission[];
+    const checkResults = await checkMultiple(permissionKeys);
 
-  checkAndRequest = async () => {
-    const checkResult = await this.check();
-
-    if (checkResult === RESULTS.DENIED) {
-      await this.request();
-    }
+    const requestKeys = permissionKeys.filter(p => checkResults[p] === RESULTS.DENIED);
+    await requestMultiple(requestKeys);
+  } catch (error) {
+    console.error(error);
   }
-  
-  check = async () => this.key && await check(this.key);
-  request = async () => this.key && await request(this.key);
-  getKey = (os: OS): Permission|undefined => undefined;
+};
+
+const PermissionFactory = (type: string) => {
+  switch (type) {
+    case Permissions.Camera:
+      return getCameraPermission();
+    case Permissions.PhotoLibrary:
+      return getPhotoLibraryPermissionKey();
+    default:
+      return undefined;
+  }
 }
+
+const getCameraPermission = (): Permission => {
+  switch (Platform.OS) {
+    case OS.IOS:
+      return PERMISSIONS.IOS.CAMERA;
+    case OS.Android:
+      return PERMISSIONS.ANDROID.CAMERA;
+    default:
+      throw Error("os is not defined");
+  }
+};
+
+const getPhotoLibraryPermissionKey = (): Permission => {
+  switch (Platform.OS) {
+    case OS.IOS:
+      return PERMISSIONS.IOS.PHOTO_LIBRARY;
+    case OS.Android:
+      return PERMISSIONS.ANDROID.READ_MEDIA_IMAGES;
+    default:
+      throw Error("os is not defined");
+  }
+};
